Initial version of WhileyWeb.rs
This is an initial version of WhileyWeb.rs which is uses Nickel.rs as
the underlying web framework.  At this stage, it doesn't do much except
serve up the index page and provide access to assets (e.g.
js,css,templates, etc).

The major piece of the jigsaw to figure out next will be executing the
JVM from within rust in order to run the Whiley compiler and/or
generator Whiley programs.